requirements.txt
azure-core==1.38.2
azure-storage-blob==12.28.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
idna==3.11
isodate==0.7.2
pycparser==3.0
python-dotenv==1.2.2
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.3
